pyproject.toml
[project]
name = "insights"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<=4.0.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "openai (>=1.77.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "azure-devops (>=7.1.0b4,<8.0.0)",
    "azure-identity (>=1.21.0,<2.0.0)",
    "azure-kusto-data (>=5.0.3,<6.0.0)"
]

[tool.poetry]
packages = [{include = "adokqlchatbot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
